create parser.section.offense
